refactor(core): move elevation data process in xbilparser
